tx: remove static constructors
