+++ packages/tx/src/1559/constructors.ts
@@ -1,0 +1,103 @@
+import { RLP } from '@ethereumjs/rlp'
+import { bytesToBigInt, bytesToHex, equalsBytes, validateNoLeadingZeroes } from '@ethereumjs/util'
+
+import { TransactionType } from '../types.js'
+import { txTypeBytes, validateNotArray } from '../util.js'
+
+import { FeeMarketEIP1559Transaction } from './tx.js'
+
+import type { TxOptions } from '../types.js'
+import type { TxData, TxValuesArray } from './tx.js'
+
+/**
+ * Instantiate a transaction from a data dictionary.
+ *
+ * Format: { chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, v, r, s }
+ *
+ * Notes:
+ * - `chainId` will be set automatically if not provided
+ * - All parameters are optional and have some basic default values
+ */
+export function create1559FeeMarketTx(txData: TxData, opts: TxOptions = {}) {
+  return new FeeMarketEIP1559Transaction(txData, opts)
+}
+
+/**
+ * Create a transaction from a values array.
+ *
+ * Format: `[chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, signatureYParity, signatureR, signatureS]`
+ */
+export function createEIP1559FeeMarketTxFromBytesArray(
+  values: TxValuesArray,
+  opts: TxOptions = {}
+) {
+  if (values.length !== 9 && values.length !== 12) {
+    throw new Error(
+      'Invalid EIP-1559 transaction. Only expecting 9 values (for unsigned tx) or 12 values (for signed tx).'
+    )
+  }
+
+  const [
+    chainId,
+    nonce,
+    maxPriorityFeePerGas,
+    maxFeePerGas,
+    gasLimit,
+    to,
+    value,
+    data,
+    accessList,
+    v,
+    r,
+    s,
+  ] = values
+
+  validateNotArray({ chainId, v })
+  validateNoLeadingZeroes({ nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, value, v, r, s })
+
+  return new FeeMarketEIP1559Transaction(
+    {
+      chainId: bytesToBigInt(chainId),
+      nonce,
+      maxPriorityFeePerGas,
+      maxFeePerGas,
+      gasLimit,
+      to,
+      value,
+      data,
+      accessList: accessList ?? [],
+      v: v !== undefined ? bytesToBigInt(v) : undefined, // EIP2930 supports v's with value 0 (empty Uint8Array)
+      r,
+      s,
+    },
+    opts
+  )
+}
+
+/**
+ * Instantiate a transaction from the serialized tx.
+ *
+ * Format: `0x02 || rlp([chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, signatureYParity, signatureR, signatureS])`
+ */
+export function create1559FeeMarketTxFromRLP(serialized: Uint8Array, opts: TxOptions = {}) {
+  if (
+    equalsBytes(serialized.subarray(0, 1), txTypeBytes(TransactionType.FeeMarketEIP1559)) === false
+  ) {
+    throw new Error(
+      `Invalid serialized tx input: not an EIP-1559 transaction (wrong tx type, expected: ${
+        TransactionType.FeeMarketEIP1559
+      }, received: ${bytesToHex(serialized.subarray(0, 1))}`
+    )
+  }
+
+  const values = RLP.decode(serialized.subarray(1))
+
+  if (!Array.isArray(values)) {
+    throw new Error('Invalid serialized tx input: must be array')
+  }
+
+  return createEIP1559FeeMarketTxFromBytesArray(values as TxValuesArray, opts)
+}
