+++ packages/tx/src/1559/index.ts
@@ -1,0 +1,2 @@
+export * from './constructors.js'
+export { FeeMarketEIP1559Transaction } from './tx.js'
