+++ packages/tx/src/2930/index.ts
@@ -1,0 +1,2 @@
+export * from './constructors.js'
+export { AccessListEIP2930Transaction } from './tx.js'
