+++ packages/tx/src/2930/tx.ts
@@ -1,14 +1,10 @@
-import { RLP } from '@ethereumjs/rlp'
 import {
   BIGINT_27,
   MAX_INTEGER,
   bigIntToHex,
   bigIntToUnpaddedBytes,
   bytesToBigInt,
-  bytesToHex,
-  equalsBytes,
   toBytes,
-  validateNoLeadingZeroes,
 } from '@ethereumjs/util'
 
 import { BaseTransaction } from '../baseTransaction.js'
@@ -16,7 +12,9 @@ import * as EIP2718 from '../capabilities/eip2718.js'
 import * as EIP2930 from '../capabilities/eip2930.js'
 import * as Legacy from '../capabilities/legacy.js'
 import { TransactionType } from '../types.js'
-import { AccessLists, txTypeBytes } from '../util.js'
+import { AccessLists, validateNotArray } from '../util.js'
+
+import { create2930AccessListTx } from './constructors.js'
 
 import type {
   AccessList,
@@ -28,8 +26,8 @@ import type {
 } from '../types.js'
 import type { Common } from '@ethereumjs/common'
 
-type TxData = AllTypesTxData[TransactionType.AccessListEIP2930]
-type TxValuesArray = AllTypesTxValuesArray[TransactionType.AccessListEIP2930]
+export type TxData = AllTypesTxData[TransactionType.AccessListEIP2930]
+export type TxValuesArray = AllTypesTxValuesArray[TransactionType.AccessListEIP2930]
 
 /**
  * Typed transaction with optional access lists
@@ -44,85 +42,6 @@ export class AccessListEIP2930Transaction extends BaseTransaction<TransactionTyp
   public readonly gasPrice: bigint
 
   public readonly common: Common
-
-  /**
-   * Instantiate a transaction from a data dictionary.
-   *
-   * Format: { chainId, nonce, gasPrice, gasLimit, to, value, data, accessList,
-   * v, r, s }
-   *
-   * Notes:
-   * - `chainId` will be set automatically if not provided
-   * - All parameters are optional and have some basic default values
-   */
-  public static fromTxData(txData: TxData, opts: TxOptions = {}) {
-    return new AccessListEIP2930Transaction(txData, opts)
-  }
-
-  /**
-   * Instantiate a transaction from the serialized tx.
-   *
-   * Format: `0x01 || rlp([chainId, nonce, gasPrice, gasLimit, to, value, data, accessList,
-   * signatureYParity (v), signatureR (r), signatureS (s)])`
-   */
-  public static fromSerializedTx(serialized: Uint8Array, opts: TxOptions = {}) {
-    if (
-      equalsBytes(serialized.subarray(0, 1), txTypeBytes(TransactionType.AccessListEIP2930)) ===
-      false
-    ) {
-      throw new Error(
-        `Invalid serialized tx input: not an EIP-2930 transaction (wrong tx type, expected: ${
-          TransactionType.AccessListEIP2930
-        }, received: ${bytesToHex(serialized.subarray(0, 1))}`
-      )
-    }
-
-    const values = RLP.decode(Uint8Array.from(serialized.subarray(1)))
-
-    if (!Array.isArray(values)) {
-      throw new Error('Invalid serialized tx input: must be array')
-    }
-
-    return AccessListEIP2930Transaction.fromValuesArray(values as TxValuesArray, opts)
-  }
-
-  /**
-   * Create a transaction from a values array.
-   *
-   * Format: `[chainId, nonce, gasPrice, gasLimit, to, value, data, accessList,
-   * signatureYParity (v), signatureR (r), signatureS (s)]`
-   */
-  public static fromValuesArray(values: TxValuesArray, opts: TxOptions = {}) {
-    if (values.length !== 8 && values.length !== 11) {
-      throw new Error(
-        'Invalid EIP-2930 transaction. Only expecting 8 values (for unsigned tx) or 11 values (for signed tx).'
-      )
-    }
-
-    const [chainId, nonce, gasPrice, gasLimit, to, value, data, accessList, v, r, s] = values
-
-    this._validateNotArray({ chainId, v })
-    validateNoLeadingZeroes({ nonce, gasPrice, gasLimit, value, v, r, s })
-
-    const emptyAccessList: AccessList = []
-
-    return new AccessListEIP2930Transaction(
-      {
-        chainId: bytesToBigInt(chainId),
-        nonce,
-        gasPrice,
-        gasLimit,
-        to,
-        value,
-        data,
-        accessList: accessList ?? emptyAccessList,
-        v: v !== undefined ? bytesToBigInt(v) : undefined, // EIP2930 supports v's with value 0 (empty Uint8Array)
-        r,
-        s,
-      },
-      opts
-    )
-  }
 
   /**
    * This constructor takes the values, validates them, assigns them and freezes the object.
@@ -157,7 +76,7 @@ export class AccessListEIP2930Transaction extends BaseTransaction<TransactionTyp
       gasPrice: this.gasPrice,
     })
 
-    BaseTransaction._validateNotArray(txData)
+    validateNotArray(txData)
 
     if (this.gasPrice * this.gasLimit > MAX_INTEGER) {
       const msg = this._errorMsg('gasLimit * gasPrice cannot exceed MAX_INTEGER')
@@ -294,7 +213,7 @@ export class AccessListEIP2930Transaction extends BaseTransaction<TransactionTyp
     s = toBytes(s)
     const opts = { ...this.txOptions, common: this.common }
 
-    return AccessListEIP2930Transaction.fromTxData(
+    return create2930AccessListTx(
       {
         chainId: this.chainId,
         nonce: this.nonce,
