+++ packages/tx/src/4844/index.ts
@@ -1,0 +1,2 @@
+export * from './constructors.js'
+export { BlobEIP4844Transaction } from './tx.js'
