+++ packages/tx/src/7702/constructors.ts
@@ -1,0 +1,102 @@
+import { RLP } from '@ethereumjs/rlp'
+import { bytesToBigInt, bytesToHex, equalsBytes, validateNoLeadingZeroes } from '@ethereumjs/util'
+
+import { TransactionType } from '../types.js'
+import { txTypeBytes, validateNotArray } from '../util.js'
+
+import { EOACodeEIP7702Transaction } from './tx.js'
+
+import type { TxOptions } from '../types.js'
+import type { TxData, TxValuesArray } from './tx.js'
+
+/**
+ * Instantiate a transaction from a data dictionary.
+ *
+ * Format: { chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, v, r, s }
+ *
+ * Notes:
+ * - `chainId` will be set automatically if not provided
+ * - All parameters are optional and have some basic default values
+ */
+export function create7702EOACodeTx(txData: TxData, opts: TxOptions = {}) {
+  return new EOACodeEIP7702Transaction(txData, opts)
+}
+
+/**
+ * Create a transaction from a values array.
+ *
+ * Format: `[chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, signatureYParity, signatureR, signatureS]`
+ */
+export function create7702EOACodeTxFromBytesArray(values: TxValuesArray, opts: TxOptions = {}) {
+  if (values.length !== 10 && values.length !== 13) {
+    throw new Error(
+      'Invalid EIP-7702 transaction. Only expecting 10 values (for unsigned tx) or 13 values (for signed tx).'
+    )
+  }
+
+  const [
+    chainId,
+    nonce,
+    maxPriorityFeePerGas,
+    maxFeePerGas,
+    gasLimit,
+    to,
+    value,
+    data,
+    accessList,
+    authorityList,
+    v,
+    r,
+    s,
+  ] = values
+
+  validateNotArray({ chainId, v })
+  validateNoLeadingZeroes({ nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, value, v, r, s })
+
+  return new EOACodeEIP7702Transaction(
+    {
+      chainId: bytesToBigInt(chainId),
+      nonce,
+      maxPriorityFeePerGas,
+      maxFeePerGas,
+      gasLimit,
+      to,
+      value,
+      data,
+      accessList: accessList ?? [],
+      authorizationList: authorityList ?? [],
+      v: v !== undefined ? bytesToBigInt(v) : undefined, // EIP2930 supports v's with value 0 (empty Uint8Array)
+      r,
+      s,
+    },
+    opts
+  )
+}
+
+/**
+ * Instantiate a transaction from the serialized tx.
+ *
+ * Format: `0x04 || rlp([chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gasLimit, to, value, data,
+ * accessList, signatureYParity, signatureR, signatureS])`
+ */
+export function create7702EOACodeTxFromRLP(serialized: Uint8Array, opts: TxOptions = {}) {
+  if (
+    equalsBytes(serialized.subarray(0, 1), txTypeBytes(TransactionType.EOACodeEIP7702)) === false
+  ) {
+    throw new Error(
+      `Invalid serialized tx input: not an EIP-7702 transaction (wrong tx type, expected: ${
+        TransactionType.EOACodeEIP7702
+      }, received: ${bytesToHex(serialized.subarray(0, 1))}`
+    )
+  }
+
+  const values = RLP.decode(serialized.subarray(1))
+
+  if (!Array.isArray(values)) {
+    throw new Error('Invalid serialized tx input: must be array')
+  }
+
+  return create7702EOACodeTxFromBytesArray(values as TxValuesArray, opts)
+}
