+++ packages/tx/src/7702/index.ts
@@ -1,0 +1,2 @@
+export * from './constructors.js'
+export { EOACodeEIP7702Transaction } from './tx.js'
