+++ packages/tx/src/baseTransaction.ts
@@ -464,31 +464,6 @@ export abstract class BaseTransaction<T extends TransactionType>
     }
   }
 
-  protected static _validateNotArray(values: { [key: string]: any }) {
-    const txDataKeys = [
-      'nonce',
-      'gasPrice',
-      'gasLimit',
-      'to',
-      'value',
-      'data',
-      'v',
-      'r',
-      's',
-      'type',
-      'baseFee',
-      'maxFeePerGas',
-      'chainId',
-    ]
-    for (const [key, value] of Object.entries(values)) {
-      if (txDataKeys.includes(key)) {
-        if (Array.isArray(value)) {
-          throw new Error(`${key} cannot be an array`)
-        }
-      }
-    }
-  }
-
   /**
    * Return a compact error string representation of the object
    */
