+++ packages/tx/src/legacy/index.ts
@@ -1,0 +1,2 @@
+export * from './constructors.js'
+export { LegacyTransaction } from './tx.js'
