+++ packages/tx/src/util.ts
@@ -223,3 +223,28 @@ export class AuthorizationLists {
 export function txTypeBytes(txType: TransactionType): Uint8Array {
   return hexToBytes(`0x${txType.toString(16).padStart(2, '0')}`)
 }
+
+export function validateNotArray(values: { [key: string]: any }) {
+  const txDataKeys = [
+    'nonce',
+    'gasPrice',
+    'gasLimit',
+    'to',
+    'value',
+    'data',
+    'v',
+    'r',
+    's',
+    'type',
+    'baseFee',
+    'maxFeePerGas',
+    'chainId',
+  ]
+  for (const [key, value] of Object.entries(values)) {
+    if (txDataKeys.includes(key)) {
+      if (Array.isArray(value)) {
+        throw new Error(`${key} cannot be an array`)
+      }
+    }
+  }
+}
